Provide basic GNOME integration
This commit integrates ``hamster-gtk`` as a integrated GNOME
application. In order to do so, a dedicated ``application__id`` is
assigned (which allows ``Gtk.Application`` to manage instance
uniqueness). Furthermore we now provide a ``*.desktop`` file with basic
metadata that is used to 'register' the app with GNOME.
To make this process as easy as possible, we also provide a new make
tarket ``register-gtk`` which uses the ``desktop-file-install`` helper
tool to handle copying the desktop file to the appropiate location.

On top of those changes this commit also makes sure that ``make
install`` will build a ``gresources`` file beforehand.

Closes: #197
Closes: #199