Introduce docker based unittest setup
This commit finally manages to establish a proper unittesting setup.

When trying to get our test setup to run with travis, several things get
in our way:
1. xvfb needs to be run not with ``xvfb-run`` but as per [these
instructions](https://docs.travis-ci.com/user/gui-and-headless-browsers/#Using-xvfb-to-Run-Tests-That-Require-a-GUI)
2. We can not actually use any of the travis build environmens for our
puposes as we target GTK>=3.10 which is unavailable in ``precise`` and
``trusty``.

As per ebassis advice, the way to deal with this is running the tests
within a docker container, which is what this PR does:
1. We provide  a new docker file that if used to build a container
provides us with a debian squeeze based environment that features
GTK>=3.10 as well as all the packages we need for testing.
This container will also clone the ``hamster-gtk`` repository and run
install the package (within the default branch, which should be
``devel``).
2. We provide a new helper script file ``run_docker.sh`` that contains
the actuall test instructions to be executed within the docker
container. This mainly means setting up xvfb and running the appropiate
make targets. Once tests have been finished it triggers the codecov API
to upload coverage reports.
3. travis config is adjusted to facilitate the new docker infrastructure
and is generally slimed down as most of the actual testing is now done
in the docker setup rather than within the travis config.
All travis does now is to build the docker container and run it
executing the new helper script.
One thing to note here is that we trigger the codecov-API as [per
instructions](https://docs.codecov.io/docs/testing-with-docker) to pass
some additional environment variables to the container

While not perfect yet, this setup should finally give us a working
framework to actually test the code (before we just ran linters and
such) on every commit/PR greatly increasing our confidence in the
codebase.